conversion to typescript
